ai_features/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
transformers==4.35.2
torch==2.2.2
web3==6.11.1
python-dotenv==1.0.0
ipfshttpclient==0.7.0
pillow==10.1.0
numpy==1.26.2
pandas==2.1.3
pydantic==2.5.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
PyMuPDF==1.23.26
sentencepiece
textblob==0.17.1
nltk==3.8.1
scikit-learn==1.3.2